+++ Cargo.toml
@@ -17,6 +17,7 @@ members = [
     "src/catalog",
     "src/catalog-debug",
     "src/catalog-protos",
+    "src/catalog-types",
     "src/ccsr",
     "src/cloud-api",
     "src/cloud-provider",
@@ -142,6 +143,7 @@ default-members = [
     "src/catalog",
     "src/catalog-debug",
     "src/catalog-protos",
+    "src/catalog-types",
     "src/ccsr",
     "src/cloud-api",
     "src/cloud-provider",
