+++ src/catalog-protos/src/serialization.rs
@@ -15,13 +15,14 @@
 
 use std::time::Duration;
 
-use mz_compute_types::config::ComputeReplicaLogging;
-use mz_controller_types::ReplicaId;
+use mz_catalog_types::cluster::ReplicaId;
+use mz_catalog_types::cluster::StorageInstanceId;
+use mz_catalog_types::compute::ComputeReplicaLogging;
 use mz_proto::{IntoRustIfSome, ProtoMapEntry, ProtoType, RustType, TryFromProtoError};
 use mz_repr::adt::mz_acl_item::{AclMode, MzAclItem};
 use mz_repr::network_policy_id::NetworkPolicyId;
 use mz_repr::role_id::RoleId;
-use mz_repr::{CatalogItemId, GlobalId, RelationVersion, Timestamp};
+use mz_repr::{CatalogItemId, GlobalId, RelationVersion};
 use mz_sql::catalog::{CatalogItemType, ObjectType, RoleAttributes, RoleMembership, RoleVars};
 use mz_sql::names::{
     CommentObjectId, DatabaseId, ResolvedDatabaseSpecifier, SchemaId, SchemaSpecifier,
@@ -31,7 +32,6 @@ use mz_sql::plan::{
     PolicyAddress,
 };
 use mz_sql::session::vars::OwnedVarInput;
-use mz_storage_types::instances::StorageInstanceId;
 
 use crate::objects::Empty;
 
@@ -563,18 +563,6 @@ impl RustType<crate::objects::EpochMillis> for u64 {
 
     fn from_proto(proto: crate::objects::EpochMillis) -> Result<Self, TryFromProtoError> {
         Ok(proto.millis)
-    }
-}
-
-impl RustType<crate::objects::Timestamp> for Timestamp {
-    fn into_proto(&self) -> crate::objects::Timestamp {
-        crate::objects::Timestamp {
-            internal: self.into(),
-        }
-    }
-
-    fn from_proto(proto: crate::objects::Timestamp) -> Result<Self, TryFromProtoError> {
-        Ok(Timestamp::new(proto.internal))
     }
 }
