+++ src/catalog-types/Cargo.toml
@@ -1,0 +1,27 @@
+[package]
+name = "mz-catalog-types"
+version = "0.0.0"
+edition.workspace = true
+rust-version.workspace = true
+publish = false
+
+[lints]
+workspace = true
+
+# The crate deliberately has few dependencies. Please don't add new dependencies without
+# discussing with the Materialize team. Specifically, avoid adding dependencies that
+# bring in large dependency trees.
+[dependencies]
+serde = { version = "1.0.219", features = ["derive"] }
+tracing = "0.1.37"
+workspace-hack = { version = "0.0.0", path = "../workspace-hack", optional = true }
+
+[dev-dependencies]
+
+[build-dependencies]
+
+[package.metadata.cargo-udeps.ignore]
+normal = ["workspace-hack"]
+
+[features]
+default = ["workspace-hack"]
