+++ src/catalog-types/src/cluster.rs
@@ -1,0 +1,207 @@
+// Copyright Materialize, Inc. and contributors. All rights reserved.
+//
+// Use of this software is governed by the Business Source License
+// included in the LICENSE file.
+//
+// As of the Change Date specified in that file, in accordance with
+// the Business Source License, use of this software will be governed
+// by the Apache License, Version 2.0.
+
+//! Types related to clusters.
+
+use std::fmt;
+use std::str::FromStr;
+
+use serde::{Deserialize, Serialize};
+use tracing::error;
+
+/// Identifier of a storage instance.
+#[derive(Clone, Copy, Debug, Eq, PartialEq, Ord, PartialOrd, Hash, Serialize, Deserialize)]
+pub enum StorageInstanceId {
+    /// A system storage instance.
+    System(u64),
+    /// A user storage instance.
+    User(u64),
+}
+
+impl StorageInstanceId {
+    /// Creates a new `StorageInstanceId` in the system namespace. The top 16 bits of `id` must be
+    /// 0, because this ID is packed into 48 bits of
+    /// `mz_repr::GlobalId::IntrospectionSourceIndex`.
+    pub fn system(id: u64) -> Option<Self> {
+        Self::new(id, Self::System)
+    }
+
+    /// Creates a new `StorageInstanceId` in the user namespace. The top 16 bits of `id` must be
+    /// 0, because this ID is packed into 48 bits of
+    /// `mz_repr::GlobalId::IntrospectionSourceIndex`.
+    pub fn user(id: u64) -> Option<Self> {
+        Self::new(id, Self::User)
+    }
+
+    fn new(id: u64, variant: fn(u64) -> Self) -> Option<Self> {
+        const MASK: u64 = 0xFFFF << 48;
+        const WARN_MASK: u64 = 1 << 47;
+        if MASK & id == 0 {
+            if WARN_MASK & id != 0 {
+                error!("{WARN_MASK} or more `StorageInstanceId`s allocated, we will run out soon");
+            }
+            Some(variant(id))
+        } else {
+            None
+        }
+    }
+
+    /// Extract the inner u64 ID.
+    pub fn inner_id(&self) -> u64 {
+        match self {
+            StorageInstanceId::System(id) | StorageInstanceId::User(id) => *id,
+        }
+    }
+
+    /// Returns true if this represents a user object.
+    pub fn is_user(&self) -> bool {
+        matches!(self, Self::User(_))
+    }
+
+    /// Returns true if this represents a system object.
+    pub fn is_system(&self) -> bool {
+        matches!(self, Self::System(_))
+    }
+}
+
+impl FromStr for StorageInstanceId {
+    type Err = IdParseError<Cluster>;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        if s.len() < 2 {
+            return Err(s.into());
+        }
+        let val: u64 = s[1..].parse()?;
+        match s.chars().next().unwrap() {
+            's' => Ok(Self::System(val)),
+            'u' => Ok(Self::User(val)),
+            _ => Err(s.into()),
+        }
+    }
+}
+
+/// An error parsing a `StorageInstanceId`.
+#[derive(Debug)]
+pub struct IdParseError<V> {
+    reason: String,
+    _marker: std::marker::PhantomData<V>,
+}
+
+impl<V: fmt::Debug> std::error::Error for IdParseError<V> {}
+
+impl<V> fmt::Display for IdParseError<V> {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "couldn't parse id {}", self.reason)
+    }
+}
+
+impl<V> From<&str> for IdParseError<V> {
+    fn from(reason: &str) -> Self {
+        reason.to_string().into()
+    }
+}
+
+impl<V> From<std::num::ParseIntError> for IdParseError<V> {
+    fn from(error: std::num::ParseIntError) -> Self {
+        error.to_string().into()
+    }
+}
+
+impl<V> From<String> for IdParseError<V> {
+    fn from(reason: String) -> Self {
+        Self {
+            reason,
+            _marker: std::marker::PhantomData,
+        }
+    }
+}
+
+impl fmt::Display for StorageInstanceId {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        match self {
+            Self::System(id) => write!(f, "s{}", id),
+            Self::User(id) => write!(f, "u{}", id),
+        }
+    }
+}
+
+/// Identifier of a replica.
+#[derive(Clone, Copy, Debug, Eq, PartialEq, Ord, PartialOrd, Hash, Serialize, Deserialize)]
+pub enum ReplicaId {
+    /// A user replica.
+    User(u64),
+    /// A system replica.
+    System(u64),
+}
+
+impl ReplicaId {
+    /// Return the inner numeric ID value.
+    pub fn inner_id(&self) -> u64 {
+        match self {
+            ReplicaId::User(id) => *id,
+            ReplicaId::System(id) => *id,
+        }
+    }
+
+    /// Whether this value identifies a user replica.
+    pub fn is_user(&self) -> bool {
+        matches!(self, Self::User(_))
+    }
+
+    /// Whether this value identifies a system replica.
+    pub fn is_system(&self) -> bool {
+        matches!(self, Self::System(_))
+    }
+}
+
+impl fmt::Display for ReplicaId {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        match self {
+            Self::User(id) => write!(f, "u{}", id),
+            Self::System(id) => write!(f, "s{}", id),
+        }
+    }
+}
+
+impl FromStr for ReplicaId {
+    type Err = IdParseError<Replica>;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        let first = s.chars().next();
+        let rest = s.get(1..);
+        if let (Some(prefix), Some(num)) = (first, rest) {
+            let id = num.parse()?;
+            match prefix {
+                'u' => return Ok(Self::User(id)),
+                's' => return Ok(Self::System(id)),
+                _ => (),
+            }
+        }
+
+        Err(s.into())
+    }
+}
+
+/// A marker type for replica ID parse errors.
+#[derive(Debug)]
+pub struct Replica;
+impl fmt::Display for Replica {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        write!(f, "replica")
+    }
+}
+
+/// A marker type for cluster ID parse errors.
+#[derive(Debug)]
+pub struct Cluster;
+impl fmt::Display for Cluster {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        write!(f, "cluster")
+    }
+}
