+++ src/cluster-client/Cargo.toml
@@ -10,7 +10,7 @@ publish = false
 workspace = true
 
 [dependencies]
-anyhow = "1.0.100"
+mz-catalog-types = { path = "../catalog-types" }
 mz-ore = { path = "../ore", features = ["tracing"] }
 mz-repr = { path = "../repr" }
 prometheus = { version = "0.14.0", default-features = false }
