+++ src/repr/src/network_policy_id.rs
@@ -11,8 +11,6 @@ use std::fmt;
 use std::str::FromStr;
 
 use anyhow::{Error, anyhow};
-use columnation::{Columnation, CopyRegion};
-use mz_lowertest::MzReflect;
 use proptest_derive::Arbitrary;
 use serde::{Deserialize, Serialize};
 
@@ -21,18 +19,7 @@ const USER_CHAR: char = 'u';
 
 /// The identifier for a network policy.
 #[derive(
-    Arbitrary,
-    Clone,
-    Copy,
-    Debug,
-    Eq,
-    PartialEq,
-    Ord,
-    PartialOrd,
-    Hash,
-    Serialize,
-    Deserialize,
-    MzReflect,
+    Arbitrary, Clone, Copy, Debug, Eq, PartialEq, Ord, PartialOrd, Hash, Serialize, Deserialize,
 )]
 pub enum NetworkPolicyId {
     System(u64),
@@ -87,10 +74,6 @@ impl fmt::Display for NetworkPolicyId {
             Self::User(id) => write!(f, "{USER_CHAR}{id}"),
         }
     }
-}
-
-impl Columnation for NetworkPolicyId {
-    type InnerRegion = CopyRegion<NetworkPolicyId>;
 }
 
 #[mz_ore::test]
