+++ src/storage-types/Cargo.toml
@@ -35,6 +35,7 @@ mysql_async = { version = "0.36.1", default-features = false, features = [
 ] }
 mz-aws-util = { path = "../aws-util", features = ["s3"] }
 mz-ccsr = { path = "../ccsr" }
+mz-catalog-types = { path = "../catalog-types" }
 mz-cloud-resources = { path = "../cloud-resources" }
 mz-dyncfg = { path = "../dyncfg" }
 mz-expr = { path = "../expr" }
